Refactor file organization logic to prevent overwriting existing files

+++ autoorganizer.py
@@ -15,7 +15,6 @@ def find_category(file_extension, type_map=TYPE_RULES):
         subfolders = details.get("subfolders", {})
         if subfolders:
             sub_category = find_category(file_extension, subfolders)
-            print(sub_category)
             if sub_category and sub_category != "Other":  # Only return if it's a valid match
                 return os.path.join(category, sub_category)
 
@@ -24,7 +23,6 @@ def find_category(file_extension, type_map=TYPE_RULES):
 
 def get_file_type(file_path):
     ext = os.path.splitext(file_path)[1].lower()
-    print(ext)
     return find_category(ext)
 
 
@@ -78,13 +76,16 @@ def organize_files(base_dir, config: Config):
                     folder_name = get_pattern_folder(file)
                 else:
                     folder_name = "Unsorted"
-                print(f"folder name: {folder_name} base dir: {base_dir}")
                 target_dir = os.path.join(base_dir, folder_name)
                 os.makedirs(target_dir, exist_ok=True)
 
                 new_path = os.path.join(target_dir, file)
                 if file_path != new_path:
-                    shutil.move(file_path, new_path)
-                    # print(f"Moved: {file_path} -> {new_path}")
+                    if os.path.exists(new_path):
+                        print(
+                            f"Skipped: '{file}' already exists in '{target_dir}'")
+                    else:
+                        shutil.move(file_path, new_path)
+
     except Exception as e:
         print(f"Error during file organization: {e}")
